Fix/qd 87 rsi divergence (#169)
* refactor: dir indicator

* feat: zigzag candel

* refactor: dir indicator

* refactor: rename ta
