+++ ezyquant/indicators/ta.py
@@ -1,10 +1,8 @@
-import operator
 from typing import Callable, Tuple
 
 import numpy as np
 import pandas as pd
-import zigzag
-from ta.momentum import ROCIndicator, RSIIndicator, StochasticOscillator, rsi
+from ta.momentum import ROCIndicator, RSIIndicator, StochasticOscillator
 from ta.trend import MACD, ADXIndicator, CCIIndicator, IchimokuIndicator, PSARIndicator
 from ta.utils import _ema, _sma
 from ta.volatility import (
@@ -14,7 +12,8 @@ from ta.volatility import (
     KeltnerChannel,
 )
 
-from .errors import InputError
+from ..errors import InputError
+from .rsi_divergence import rsi_divergence
 
 nan = float("nan")
 
@@ -354,6 +353,8 @@ class TA:
 
     @staticmethod
     def rsi_divergence(
+        high: pd.DataFrame,
+        low: pd.DataFrame,
         close: pd.DataFrame,
         rsi_period: int = 14,
         pivot_up_thresh: float = 0.05,
@@ -363,6 +364,10 @@ class TA:
 
         Parameters
         ----------
+        high: pd.DataFrame
+            dataset 'High' dataframe.
+        low: pd.DataFrame
+            dataset 'Low' dataframe.
         close: pd.DataFrame
             dataset 'Close' dataframe.
         rsi_period: int = 14
@@ -380,6 +385,8 @@ class TA:
         out = close.apply(
             lambda x: rsi_divergence(
                 close=x,
+                high=high[x.name],
+                low=low[x.name],
                 rsi_period=rsi_period,
                 pivot_up_thresh=pivot_up_thresh,
                 pivot_down_thresh=pivot_down_thresh,
@@ -712,110 +719,3 @@ def _apply_t(series: pd.Series, func: Callable) -> pd.DataFrame:
         raise InputError(f"{func.__name__} returned an empty dataframe")
     assert isinstance(df, pd.DataFrame)
     return df
-
-
-"""
-Pivot
-"""
-
-
-def pivot_points_high(high: pd.Series, left_bars: int, right_bars: int) -> pd.Series:
-    """Pivot Points High.
-
-    High is the highest high in a window between left_bars and
-    right_bars. If multiple bars have the same high, the most right bar
-    is used.
-    """
-    max_ = high.rolling(window=left_bars + 1).max()
-    max_right = high.rolling(window=right_bars).max().shift(-right_bars)
-
-    return high.where((max_ == high) & (max_ > max_right))
-
-
-def pivot_points_low(low: pd.Series, left_bars: int, right_bars: int) -> pd.Series:
-    """Pivot Points Low.
-
-    Low is the lowest low in a window between left_bars and right_bars.
-    If multiple bars have the same low, the most right bar is used.
-    """
-    min_ = low.rolling(window=left_bars + 1).min()
-    min_right = low.rolling(window=right_bars).min().shift(-right_bars)
-
-    return low.where((min_ == low) & (min_ < min_right))
-
-
-"""
-RSI Divergence
-"""
-
-
-def rsi_divergence(
-    close: pd.Series,
-    rsi_period: int = 14,
-    pivot_up_thresh: float = 0.05,
-    pivot_down_thresh: float = -0.05,
-) -> pd.Series:
-    """Return Positive RSI of the bullish divergence points and Negative RSI of
-    the bearish divergence points.
-
-    Parameters
-    ----------
-    high : pd.Series
-        Series of 'high' prices.
-    low : pd.Series
-        Series of 'low' prices.
-    close : pd.Series
-        Series of 'close' prices.
-    rsi_period : int, optional
-        RSI period, by default 14
-    pivot_up_thresh : int
-        Pivot up threshold, by default 0.05
-    pivot_down_thresh : int
-        Pivot down threshold, by default -0.05
-    """
-    # Calculate RSI
-    rsi_ = rsi(close, window=rsi_period)
-
-    # Calculate pivot points using zigzag
-    zz_ = zigzag.peak_valley_pivots(  # type: ignore
-        close.to_numpy(),
-        up_thresh=pivot_up_thresh,
-        down_thresh=pivot_down_thresh,
-    )
-    close_pl = close[zz_ < 0]
-    close_ph = close[zz_ > 0]
-
-    # Calculate Divergence
-    bullish = _divergence(close_pl, rsi_, bullish=True)
-    bearish = _divergence(close_ph, rsi_, bullish=False)
-
-    return bullish.fillna(-bearish)
-
-
-def _divergence(
-    price_pivot: pd.Series, indicator: pd.Series, bullish: bool = True
-) -> pd.Series:
-    """Divergence.
-
-    Parameters
-    ----------
-    price_pivot : pd.Series
-        Price pivot points. low for bullish divergence and high for bearish divergence.
-    indicator : pd.Series
-        indicator series.
-    bullish : bool, optional
-        bullish or bearish divergence, by default True
-
-    Returns
-    -------
-    pd.Series
-        Divergence series. Not NaN when divergence occurs.
-    """
-    indicator_pivot = indicator[price_pivot.index]
-
-    op = operator.lt if bullish else operator.gt
-
-    return indicator.where(
-        op(price_pivot, price_pivot.shift(1))
-        & op(indicator_pivot.shift(1), indicator_pivot)
-    )
