+++ requirements-dev.txt
@@ -5,7 +5,6 @@ ta~=0.10
 XlsxWriter~=3.0
 typing_extensions~=4.4
 quantstats~=0.0
-zigzag==0.2.2 # TODO: use pandas_ta instead when it's released
 
 # Sytle
 isort~=5.0
