+++ tests/test_indicators/test_ta.py
@@ -1,14 +1,10 @@
 import string
-from functools import lru_cache
 
 import numpy as np
 import pandas as pd
 import pytest
-import yfinance as yf
-from pandas.testing import assert_series_equal
 
 from ezyquant import SETSignalCreator
-from ezyquant import indicators as ind
 from ezyquant import validators as vld
 from ezyquant.errors import InputError
 
@@ -23,13 +19,6 @@ def make_random_df(n_row: int = 1, n_col: int = 1):
         index=pd.date_range(start="2000-01-01", periods=n_row),
     )
     df[df < 0.5] = np.nan
-    return df
-
-
-@lru_cache
-def make_aapl_df():
-    df = yf.download("AAPL", start="2020-01-01", end="2020-12-31", progress=False)
-    df["Close"] = df["Adj Close"]
     return df
 
 
@@ -147,9 +136,11 @@ class TestTa:
     def test_rsi_divergence(self, n_row: int, n_col: int):
         # Mock
         df1 = make_random_df(n_row=n_row, n_col=n_col)
+        df2 = make_random_df(n_row=n_row, n_col=n_col)
+        df3 = make_random_df(n_row=n_row, n_col=n_col)
 
         # Test
-        result = SETSignalCreator.ta.rsi_divergence(df1)
+        result = SETSignalCreator.ta.rsi_divergence(df1, df2, df3)
 
         # Check
         self._check(result)
@@ -373,97 +364,3 @@ class TestTaEmpty:
         # Test
         with pytest.raises(InputError):
             SETSignalCreator.ta.kc(df, df, df)
-
-
-@pytest.mark.parametrize(
-    "data, expected",
-    [
-        # No data
-        (series_empty, series_empty),
-        # One data
-        (pd.Series([1.0]), series_empty),
-        # Data increasing
-        (pd.Series([1.0, 2.0, 3.0, 4.0, 5.0]), series_empty),
-        (pd.Series([1.0, 1.0, 2.0, 2.0, 3.0, 3.0]), series_empty),
-        # Data decreasing
-        (pd.Series([5.0, 4.0, 3.0, 2.0, 1.0]), series_empty),
-        (pd.Series([3.0, 3.0, 2.0, 2.0, 1.0, 1.0]), series_empty),
-        # Pivot edge
-        (pd.Series([1.0, 2.0, 3.0, 4.0, 3.0]), series_empty),
-        (pd.Series([3.0, 4.0, 3.0, 2.0, 1.0]), series_empty),
-        # Data same
-        (pd.Series([1.0, 1.0, 1.0, 1.0, 1.0]), series_empty),
-        # Pivot low
-        (pd.Series([3.0, 2.0, 1.0, 2.0, 3.0]), series_empty),
-        # Pivot high
-        (pd.Series([1.0, 2.0, 3.0, 2.0, 1.0]), pd.Series({2: 3.0})),
-        # Two same pivots
-        (pd.Series([1.0, 1.0, 2.0, 2.0, 1.0, 1.0]), pd.Series({3: 2.0})),
-        (pd.Series([1.0, 1.0, 2.0, 1.0, 2.0, 1.0, 1.0]), pd.Series({4: 2.0})),
-        # Two pivots high
-        (
-            pd.Series([1.0, 2.0, 3.0, 2.0, 1.0, 2.0, 3.0, 2.0, 1.0]),
-            pd.Series({2: 3.0, 6: 3.0}),
-        ),
-    ],
-)
-def test_pivot_points_high(data, expected):
-    # Test
-    actual = ind.pivot_points_high(data, 2, 2)
-
-    # Check
-    assert_series_equal(actual.dropna(), expected)
-
-
-def test_rsi_divergence():
-    # Mock
-    data = make_aapl_df()
-
-    # Test
-    actual = ind.rsi_divergence(close=data["Close"])
-
-    # Check
-    print(actual.dropna())
-
-
-@pytest.mark.parametrize(
-    "price_pivot, indicator, expected",
-    [
-        # No data
-        (series_empty, series_empty, series_empty),
-        # No divergence
-        (
-            pd.Series([1, 2, 3]),
-            pd.Series([1, 2, 3]),
-            pd.Series([nan, nan, nan]),
-        ),
-        (
-            pd.Series([3, 2, 1]),
-            pd.Series([3, 2, 1]),
-            pd.Series([nan, nan, nan]),
-        ),
-        (
-            pd.Series([1, 2]),
-            pd.Series([2, 1]),
-            pd.Series([nan, nan]),
-        ),
-        # Divergence
-        (
-            pd.Series([2, 1]),
-            pd.Series([1, 2]),
-            pd.Series([nan, 2]),
-        ),
-        # Divergence with skip price
-        (
-            pd.Series({0: 2, 2: 1}),
-            pd.Series([1, 99, 2]),
-            pd.Series([nan, nan, 2]),
-        ),
-    ],
-)
-def test_divergence(price_pivot: pd.Series, indicator: pd.Series, expected: pd.Series):
-    # Test
-    actual = ind._divergence(price_pivot=price_pivot, indicator=indicator, bullish=True)
-
-    # Check
-    assert_series_equal(actual, expected)
